# backend/src/main/resources/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "Documentation for CleanUP REST API",
    "version": "1.0.0",
    "title": "CleanUp system API"
  },
  "host": "localhost:8080",
  "basePath": "/",
...
  "paths": {
    "/auth/login": {
      "post": {
        "tags": [
          "Authorization"
        ],
        "summary": "Performs user login to the system",
        "operationId": "loginUser",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "loginDto",
            "description": "loginDto",
            "required": true,
            "schema": {
              "$ref": "#/definitions/LoginDto"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "object"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "security": [
          {
            "JWT": [
              "global"
            ]
          }
        ],
        "deprecated": false
      }
    },
...
  },
  "securityDefinitions": {
    "JWT": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header"
    }
  },
  "definitions": {
...
    "LoginDto": {
      "type": "object",
      "properties": {
        "email": {
          "type": "string"
        },
        "password": {
          "type": "string"
        }
      },
      "title": "LoginDto"
    },
...
  }
}